+++ lib/spear/connection.ex
@@ -40,6 +40,38 @@ defmodule Spear.Connection do
   See the [Security guide](guides/security.md) for information about
   certificates, credentials, and access control lists (ACLs).
 
+  ## Keep-alive
+
+  Spear and the EventStoreDB use gRPC keep-alive to ensure that any hung
+  connections are noticed and restarted.
+
+  EventStoreDB has its own configuration for keep-alive and each Spear
+  client's configuration should not necessarily match the server configuration.
+  With a `:keep_alive_interval` too low on the Spear-side and with very many
+  connected clients, the keep-alive pings can effectively become a DDoS
+  even while no clients perform any operations. This does not necessarily
+  apply to the keep-alive settings in EventStoreDB: a client connects to a
+  single EventStoreDB but an EventStoreDB may have hundreds or more clients
+  connected at once.
+
+  `:keep_alive_interval` does not express an interval in the way of
+  `:timer.send_interval/3`. Instead the keep-alive timer is optimized to
+  reset upon any data received from the connection.
+
+  This means that (assuming consistent network, which is generous) either the
+  Spear client or EventStoreDB server will dominate the keep-alive ping
+  communication, the driver of the conversation being which ever has the
+  lowest keep-alive interval configured. For this reason, it is generally
+  advisable to set the client keep-alive just higher than the server
+  keep-alive, adding noise for network latency, since the client's keep-alive
+  routine will only trigger when the server's keep-alive message is behind
+  schedule.
+
+  Note that there may not be much value in attempting to optimize the
+  keep-alive settings unless the network is very unstable: keep-alive only
+  has utility when the client, server, or network is seriously delayed or
+  silently severed.
+
   ## Examples
 
       iex> {:ok, conn} = Spear.Connection.start_link(connection_string: "esdb://localhost:2113")
@@ -47,20 +79,20 @@ defmodule Spear.Connection do
       [%Spear.Event{}, %Spear.Event{}, %Spear.Event{}]
   """
 
-  # see the very similar original implementation of this from the Mint
-  # documentation:
+  # see the very similar original implementation of this process architecture
+  # from the Mint documentation:
   # https://github.com/elixir-mint/mint/blob/796b8db097d69ede7163acba223ab2045c2773a4/pages/Architecture.md
 
   use Connection
   require Logger
 
-  alias Spear.Connection.Request
+  alias Spear.Connection.{Request, KeepAliveTimer}
   alias Spear.Connection.Configuration, as: Config
 
   @post "POST"
   @closed %Mint.TransportError{reason: :closed}
 
-  defstruct [:config, :conn, requests: %{}]
+  defstruct [:config, :conn, requests: %{}, keep_alive_timer: %KeepAliveTimer{}]
 
   @typedoc """
   A connection process
@@ -143,8 +175,11 @@ defmodule Spear.Connection do
   @impl Connection
   def connect(_, s) do
     case do_connect(s.config) do
-      {:ok, conn} -> {:ok, %__MODULE__{s | conn: conn}}
-      {:error, _reason} -> {:backoff, 500, s}
+      {:ok, conn} ->
+        {:ok, %__MODULE__{s | conn: conn, keep_alive_timer: KeepAliveTimer.start(s.config)}}
+
+      {:error, _reason} ->
+        {:backoff, 500, s}
     end
   end
 
@@ -154,7 +189,12 @@ defmodule Spear.Connection do
 
     :ok = close_requests(s)
 
-    s = %__MODULE__{s | conn: nil, requests: %{}}
+    s = %__MODULE__{
+      s
+      | conn: nil,
+        requests: %{},
+        keep_alive_timer: KeepAliveTimer.clear(s.keep_alive_timer)
+    }
 
     case info do
       {:close, from} ->
@@ -162,11 +202,8 @@ defmodule Spear.Connection do
 
         {:noconnect, s}
 
-      # coveralls-ignore-start
-      :closed ->
-        :ok
+      _ ->
         {:connect, :reconnect, s}
-        # coveralls-ignore-stop
     end
   end
 
@@ -254,6 +291,25 @@ defmodule Spear.Connection do
     end
   end
 
+  def handle_info(:keep_alive, s) do
+    case Mint.HTTP2.ping(s.conn) do
+      {:ok, conn, request_ref} ->
+        s = put_in(s.conn, conn)
+        s = update_in(s.keep_alive_timer, &KeepAliveTimer.start_timeout_timer(&1, request_ref))
+
+        {:noreply, s}
+
+      # coveralls-ignore-start
+      {:error, conn, reason} ->
+        s = put_in(s.conn, conn)
+
+        if reason == @closed, do: {:disconnect, :closed, s}, else: {:noreply, s}
+        # coveralls-ignore-stop
+    end
+  end
+
+  def handle_info(:keep_alive_expired, s), do: {:disconnect, :keep_alive_timeout, s}
+
   def handle_info(message, s) do
     with %Mint.HTTP2{} = conn <- s.conn,
          {:ok, conn, responses} <- Mint.HTTP2.stream(conn, message) do
@@ -276,6 +332,8 @@ defmodule Spear.Connection do
 
   @spec handle_responses(%__MODULE__{}, list()) :: %__MODULE__{}
   defp handle_responses(s, responses) do
+    s = update_in(s.keep_alive_timer, &KeepAliveTimer.reset_interval_timer/1)
+
     responses
     |> Enum.reduce(s, &process_response/2)
     |> Request.continue_requests()
@@ -300,11 +358,17 @@ defmodule Spear.Connection do
   end
 
   defp process_response({:pong, request_ref}, s) do
-    {{:ping, from}, s} = pop_in(s.requests[request_ref])
+    case pop_in(s.requests[request_ref]) do
+      {{:ping, from}, s} ->
+        # ping was initiated by a GenServer.call/3
+        Connection.reply(from, :pong)
 
-    Connection.reply(from, :pong)
+        s
 
-    s
+      {nil, s} ->
+        # ping was initiated by the keepalive timer
+        update_in(s.keep_alive_timer, &KeepAliveTimer.clear_after_timer(&1, request_ref))
+    end
   end
 
   defp process_response({:done, request_ref}, s) do