+++ CHANGELOG.md
@@ -15,6 +15,14 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
     - credentials are passed through the `:connection_string` option or
       as `:username` and `:password` options
 
+### Added
+
+- Implemented and documented keep-alive
+    - This can be configured through the `keepAliveInterval` and
+      `keepAliveTimeout` query params in `:connection_string` or by the
+      new `:keep_alive_interval` and `:keep_alive_timeout` configuration
+      options
+
 ## 0.1.4 - 2021-04-16
 
 ### Added